fixes #9 (Performance issues?)
After the refactoring (#2) the performance of the tool dropped significantly (~200% for 864 links). The 'ruby-prof' pointed out that something is wrong with the 'visited?' method where the LinkCollector checks if a link has already been visited or not. The problem lay in the storage: I was using an Array and where a lookup cost O(n).

In a Hash the lookup costs O(1) so I converted the LinkCollector.links to a Hash by using a link as a key and nil as a value. With this approach the performance problems disappear and now the tool runs even faster than before (about 50% faster).