fixes #13 (incompatible with ruby 1.8)
When a Link is created, it uses the instance_methods for figuring out the name of the attributes. In ruby19 it returns a list of symbols and in ruby18 a list of strings. Now the Link#create converts all the results of instance_methods to strings and uses strings for comparison

Set up a ruby18 based ruby build in order to know about similar problems earlier